feat(agent): 会話分析・成約分析の数値サマリーをチャットで直接返す
これまで「会話分析」「成約・効果分析」を尋ねられるとエージェントは
get_legacy_ui_link で旧UIのURLを返すだけで、数値が一度もチャットに出て
こなかった。集計本体を summaryQueries.ts に切り出し、チャットからも同じ
数値を取得できるようにする。

- src/api/admin/analytics/summaryQueries.ts を新設し、
  GET /summary と GET /conversions の集計を fetchAnalyticsSummary /
  fetchConversionSummary として抽出（クエリ・フィールド名は無改変。
  ルート側のJSONレスポンス形状は変更なし）
- Phase52h の通知トリガー(6/7/8)は routes.ts 側に残す。チャットから
  同じ集計を呼んだときに管理者通知が発火しないようにするため
- get_analytics_summary / get_conversion_summary の2ツールを追加。
  いずれも読み取り専用で period(7d/30d)のみを取る
- プランゲートは get_legacy_ui_link の analytics/conversion と同一。
  super_admin もバイパスさせない（クライアントビューはテナントに見えて
  いる状態の再現が目的のため）
- get_legacy_ui_link の説明とシステムプロンプトを更新し、この2機能では
  グラフ詳細・低評価セッション・ABテスト閲覧のときだけ旧UIへ案内させる
- /copilot-preview の REAL_TOOL_LABEL に日本語ラベルを追加
  (読み取り専用のため REAL_WRITE_TOOLS には入れない)
