+++ src/api/admin/agent/agentRoutes.ts
@@ -450,8 +450,11 @@ export function registerAdminAgentRoutes(app: Express, db: Pool): void {
         `登録してください（同じく同一ターン内での連鎖実行は避け、ユーザーの次のメッセージを待つこと）。` +
         `PDFからの知識登録は get_legacy_ui_link(feature=knowledge_pdf) で旧管理画面へ案内してください。` +
         `請求（支払い操作）、アバタースタジオ（画像/音声/性格/ライブテスト）、エスカレーションへの有人返信、` +
-        `会話セッションの削除、会話分析、成約・効果分析、テストチャット、アバター新規作成について尋ねられた場合は、` +
+        `会話セッションの削除、テストチャット、アバター新規作成について尋ねられた場合は、` +
         `チャットで実行しようとせず get_legacy_ui_link を呼び出して旧管理画面へ案内してください。` +
+        `会話分析・成約・効果分析については get_analytics_summary / get_conversion_summary で数値サマリーを` +
+        `そのままチャットで答えてください。グラフの詳細・個別の低評価セッション・ABテスト結果を見たいと言われた` +
+        `場合のみ get_legacy_ui_link(feature=analytics / conversion) で旧管理画面へ案内してください。` +
         `ユーザーが管理画面の操作を代わりにやってほしいと頼んできた場合（例:「送料表記を直して」）は、` +
         `request_sai_task が使えます。他のLLM機能と同じ従量課金が発生するため、必ず先に依頼内容を要約提示し、` +
         `同意を得たターンでのみ confirmed=true で呼び出してください。進捗は get_sai_task_status で確認できます。` +
